requirements.txt
tensorflow
keras
pandas
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
ensure==1.0.2
joblib
types-pyYAML
streamlit
tk-tools
python-whois
-e .